+++ source/renderer/app/components/staking/stake-pools/StakePools.js
@@ -1,7 +1,8 @@
 // @flow
 import React, { Component, Fragment } from 'react';
+import SVGInline from 'react-svg-inline';
 import { observer } from 'mobx-react';
-import { defineMessages, intlShape, FormattedMessage } from 'react-intl';
+import { defineMessages, intlShape } from 'react-intl';
 import classnames from 'classnames';
 import StakePoolsRanking from './StakePoolsRanking';
 import { StakePoolsList } from './StakePoolsList';
@@ -14,7 +15,13 @@ import styles from './StakePools.scss';
 import { getFilteredStakePoolsList } from './helpers';
 import { formattedNumber } from '../../../utils/formatters';
 import StakePool from '../../../domains/StakePool';
-import { IS_RANKING_DATA_AVAILABLE } from '../../../config/stakingConfig';
+import {
+  IS_RANKING_DATA_AVAILABLE,
+  SMASH_SERVER_TYPES,
+} from '../../../config/stakingConfig';
+import smashSettingsIcon from '../../../assets/images/smash-settings-ic.inline.svg';
+import tinySpinnerIcon from '../../../assets/images/spinner-tiny.inline.svg';
+import { getSmashServerNameFromUrl } from '../../../utils/staking';
 
 const messages = defineMessages({
   delegatingListTitle: {
@@ -24,19 +31,39 @@ const messages = defineMessages({
   },
   listTitle: {
     id: 'staking.stakePools.listTitle',
-    defaultMessage: '!!!Stake pools ({pools})',
+    defaultMessage: '!!!Stake pools',
     description: '"listTitle" for the Stake Pools page.',
   },
-  listTitleWithSearch: {
-    id: 'staking.stakePools.listTitleWithSearch',
-    defaultMessage: '!!!Stake pools. Search results: ({pools})',
-    description: '"listTitle" for the Stake Pools page.',
+  listTitleLoading: {
+    id: 'staking.stakePools.listTitleLoading',
+    defaultMessage: '!!!Loading stake pools',
+    description: '"listTitleLoading" for the Stake Pools page.',
+  },
+  listTitleSearch: {
+    id: 'staking.stakePools.listTitleSearch',
+    defaultMessage: '!!!Stake pools. Search results:',
+    description: '"listTitleSearch" for the Stake Pools page.',
+  },
+  listTitleStakePools: {
+    id: 'staking.stakePools.listTitleStakePools',
+    defaultMessage: '!!!({pools})',
+    description: '"listTitleStakePools" for the Stake Pools page.',
   },
   loadingStakePoolsMessage: {
     id: 'staking.stakePools.loadingStakePoolsMessage',
     defaultMessage: '!!!Loading stake pools',
     description:
       'Loading stake pool message for the Delegation center body section.',
+  },
+  moderatedBy: {
+    id: 'staking.stakePools.moderatedBy',
+    defaultMessage: '!!!Moderated by',
+    description: 'moderatedBy message for the Delegation center body section.',
+  },
+  unmoderated: {
+    id: 'staking.stakePools.unmoderated',
+    defaultMessage: '!!!Unmoderated',
+    description: 'unmoderated message for the Delegation center body section.',
   },
 });
 
@@ -56,9 +83,12 @@ type Props = {
   stake?: ?number,
   onDelegate: Function,
   isLoading: boolean,
+  isFetching: boolean,
   isRanking: boolean,
   stakePoolsDelegatingList: Array<StakePool>,
   getStakePoolById: Function,
+  onSmashSettingsClick: Function,
+  smashServerUrl: ?string,
   maxDelegationFunds: number,
 };
 
@@ -131,9 +161,12 @@ export default class StakePools extends Component<Props, State> {
       onOpenExternalLink,
       currentTheme,
       isLoading,
+      isFetching,
       isRanking,
       stakePoolsDelegatingList,
       getStakePoolById,
+      smashServerUrl,
+      onSmashSettingsClick,
       maxDelegationFunds,
     } = this.props;
     const {
@@ -154,9 +187,12 @@ export default class StakePools extends Component<Props, State> {
         IS_RANKING_DATA_AVAILABLE && stakePool.nonMyopicMemberRewards
     ).length;
 
-    const listTitleMessage = search.trim().length
-      ? messages.listTitleWithSearch
+    const listTitleMessage = isFetching
+      ? messages.listTitleLoading
       : messages.listTitle;
+
+    const listTitleSearchMessage =
+      !!search.trim().length && intl.formatMessage(messages.listTitleSearch);
 
     const loadingSpinner = (
       <LoadingSpinner
@@ -171,6 +207,28 @@ export default class StakePools extends Component<Props, State> {
       styles.component,
       isLoading ? styles.isLoading : null,
     ]);
+
+    const smashServer = smashServerUrl
+      ? getSmashServerNameFromUrl(smashServerUrl)
+      : null;
+
+    const tinyLoadingSpinner = isFetching && (
+      <SVGInline svg={tinySpinnerIcon} className={styles.tinySpinner} />
+    );
+
+    const smashSettings = (
+      <button onClick={onSmashSettingsClick} className={styles.smashSettings}>
+        <span>
+          {smashServer && smashServer !== SMASH_SERVER_TYPES.DIRECT
+            ? intl.formatMessage(messages.moderatedBy, { smashServer })
+            : intl.formatMessage(messages.unmoderated)}
+        </span>
+        <SVGInline
+          svg={smashSettingsIcon}
+          className={styles.smashSettingsIcon}
+        />
+      </button>
+    );
 
     return (
       <div className={componentClasses}>
@@ -210,12 +268,17 @@ export default class StakePools extends Component<Props, State> {
               onListView={this.handleListView}
               isListView={isListView}
               isGridView={isGridView}
+              smashServer={smashServer}
               isClearTooltipOpeningDownward
             />
             {stakePoolsDelegatingList.length > 0 && (
               <Fragment>
                 <h2 className={styles.listTitle}>
-                  {intl.formatMessage(messages.delegatingListTitle)}
+                  <span className={styles.leftContent}>
+                    <span>
+                      {intl.formatMessage(messages.delegatingListTitle)}
+                    </span>
+                  </span>
                 </h2>
                 <StakePoolsList
                   listName={STAKE_POOLS_DELEGATING_LIST}
@@ -234,12 +297,17 @@ export default class StakePools extends Component<Props, State> {
             {isListView && (
               <Fragment>
                 <h2>
-                  <FormattedMessage
-                    {...listTitleMessage}
-                    values={{
-                      pools: formattedNumber(filteredStakePoolsList.length),
-                    }}
-                  />
+                  <span className={styles.leftContent}>
+                    <span>
+                      {intl.formatMessage(listTitleMessage)}
+                      {listTitleSearchMessage}
+                      {intl.formatMessage(messages.listTitleStakePools, {
+                        pools: formattedNumber(filteredStakePoolsList.length),
+                      })}
+                    </span>
+                    {tinyLoadingSpinner}
+                  </span>
+                  {smashSettings}
                 </h2>
                 <StakePoolsTable
                   listName={SELECTED_INDEX_TABLE}
@@ -261,12 +329,17 @@ export default class StakePools extends Component<Props, State> {
             {isGridView && (
               <Fragment>
                 <h2>
-                  <FormattedMessage
-                    {...listTitleMessage}
-                    values={{
-                      pools: formattedNumber(filteredStakePoolsList.length),
-                    }}
-                  />
+                  <span className={styles.leftContent}>
+                    <span>
+                      {intl.formatMessage(listTitleMessage)}
+                      {listTitleSearchMessage}
+                      {intl.formatMessage(messages.listTitleStakePools, {
+                        pools: formattedNumber(filteredStakePoolsList.length),
+                      })}
+                    </span>
+                    {tinyLoadingSpinner}
+                  </span>
+                  {smashSettings}
                 </h2>
                 <StakePoolsList
                   showWithSelectButton